component rearrangements
