+++ src/devTools/editor/tabs/ServicesTab.tsx
@@ -20,7 +20,6 @@ import { Button, Tab, Table } from "react-bootstrap";
 import { FieldArray, useField } from "formik";
 import { ServiceDependency } from "@/core";
 import DependencyRow from "./DependencyRow";
-import { useAuthOptions } from "@/options/pages/extensionEditor/ServiceAuthSelector";
 import { head } from "lodash";
 import { ServiceDefinition } from "@/types/definitions";
 import ServiceModal from "@/components/fields/ServiceModal";
@@ -33,6 +32,7 @@ import { browser } from "webextension-polyfill-ts";
 import { useToasts } from "react-toast-notifications";
 
 import "./ServicesTab.scss";
+import { useAuthOptions } from "@/hooks/auth";
 
 function defaultOutputKey(serviceId: string): string {
   const match = PACKAGE_REGEX.exec(serviceId);
